It's in absolute shambles.
Beginning to use the pre/postinstall scripts, and removing all the self-referential explodey shit. I think this whole thing is making a shitload of loops it shouldn't, and it's absolutely a garbage mess right now. Also making it so there's no storing of params for httpOnly cookies, as that could be a vulnerability.