Looks to mostly be ready
Had to do a bunch of testing manually. This is not what I need to be spending time on, so the testing should be something I return to at a later date. I will also be using this module within my website itself, so it will be a great testing ground there.